requirements.txt
scrapy
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.4
django-crontab==0.7.1
django-common-helpers==0.9.2
django-environ==0.4.5
django-filter==2.2.0
django-heroku==0.3.1
django-rest-swagger==2.1.2
djangorestframework==3.10.2
environ==1.0
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2-binary==2.8.3
pymongo==3.9.0
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.3
